f64 wip

+++ src/lib.rs
@@ -5,6 +5,6 @@ pub mod proc_context;
 pub mod ring_buffer;
 pub mod signal;
 
-pub fn notenum_to_frequency(notenum: u32) -> f32 {
-    440.0 * 2.0f32.powf((notenum as f32 - 69.0) / 12.0)
+pub fn notenum_to_frequency(notenum: u32) -> f64 {
+    440.0 * 2.0f64.powf((notenum as f64 - 69.0) / 12.0)
 }
